gov_services/requirements.txt
# Django Framework
Django>=5.0,<6.0

# Database
psycopg2-binary>=2.9.9

# Production Server
gunicorn>=21.2.0

# Static Files
whitenoise>=6.6.0

# Database URL Parser
dj-database-url>=2.1.0

# Basic utilities (optional)
python-dotenv>=1.0.0

# ========== نظام التصنيف الذكي للبلاغات ==========

# استخراج النص من PDF
PyPDF2>=3.0.0

# استخراج النص من الصور (OCR)
pytesseract>=0.3.10
Pillow>=10.0.0

# استخراج النص من Word
python-docx>=1.0.0

# معالجة النصوص العربية (اختياري - للترقية المستقبلية)
# transformers>=4.30.0
# torch>=2.0.0
# arabert>=1.0.0
